Improve code reusability
Split the common aspects (permission requests, Bluetooth enablement, AX
Visio connection and OpenAPI Inside App start) into a single activity.
This should make it easier to copy and paste the code into a third party
application and maybe even update the code in the future.
