+++ app/src/main/res/layout/activity_connect_request_permissions_and_bluetooth.xml
@@ -2,46 +2,50 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/constraintLayoutScreen2"
+    android:id="@+id/constraintLayoutScreen1"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".ConnectActivity">
 
     <TextView
-        android:id="@+id/textViewScreen2"
+        android:id="@+id/textViewScreen1"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
-        android:text="Screen 2"
+        android:layout_marginBottom="16dp"
+        android:text="Screen 1"
         android:textSize="34sp"
-        app:layout_constraintBottom_toTopOf="@+id/textViewConnectToAXVisio"
+        app:layout_constraintBottom_toTopOf="@+id/textViewRequestPermissionsAndBluetooth"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/textViewConnectToAXVisio"
+        android:id="@+id/textViewRequestPermissionsAndBluetooth"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:gravity="center"
-        android:text="Searching for an AX Visio device in reach. When a device is found, you can touch the button to connect to it."
+        android:text="The App needs access to Bluetooth and the location permission. Please grant it. Also please enable Bluetooth."
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <Button
-        android:id="@+id/buttonConnectToAXVisio"
+        android:id="@+id/buttonRequestPermissions"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
-        android:text="Connect To AX VIsio"
+        android:text="Request Permissions"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textViewConnectToAXVisio"
+        app:layout_constraintTop_toBottomOf="@+id/textViewRequestPermissionsAndBluetooth"
         app:layout_constraintVertical_bias="0.0" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
